Move files to Home folder
